Command bar ready for prod! (almos)
